Refactored the fadeOut function to work with the current mushrooms, rather than with the starting set, so it will work if the caps functionality is ever fixed.
Reordered the remove and insert functions to be in linear order.
Tested remove, it appears to be working correctly, the insert has
a bug. The picture source replaces correctly, but adding the un-
selected class to an element doesn't make it reappear. Tried
changing the style to 'display: inline-block', which does make
things reappear, but unstyled. Played with CSS a bit, suspect the
error is in the CSS in addition to the JS. Saving progress for the
day.