Without using complicated regexes, got the leafnode checks to work by simply checking if they're in the main hash table (Eric's idea, so much simpler). The class logic is more clear again.
Added a TON of text for the descriptions of the mushrooms, but the
formatting is pretty gross. To embed tags in javascript strings
requires things like &lt; which is also pretty gross.

The save session isn't grabbing the correct lists of things --
if you click all the way through gills and then into pores, the
leafnode for gills doesn't get cleared. I'll work on that later.