Cleaned up the code a bit, reorganized to be more logical, did a few small optimizations (moving length out of for loops, etc), all while trying to fix the session data. I had been storing an array of the id's, but that's problematic because on reload it's almost impossible to figure out what element should have which class (selected / unselected) and so the entire tree structure breaks on the next mouse click.
Moved to storing the entire object, but on reload it returns an
empty array [{},{},{} ... of the correct length. I need to look
more into how jsonify works because I think the error is in trying
to save objects rather than strings (except MDN says that's possible).

Saving progress before a break.